add update password-delete user + synchronize this with keycloak

+++ src/main/java/project/learning_managment_system/learning_managment_system_dev/user_managment/Services/ManagementService/ServiceImpl/ServiceTeacher.java
@@ -1,13 +1,17 @@
 package project.learning_managment_system.learning_managment_system_dev.user_managment.Services.ManagementService.ServiceImpl;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
 import project.learning_managment_system.learning_managment_system_dev.Config.JwtExtractor;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Dto.Student_Dto;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Dto.Teacher_Dto;
+import project.learning_managment_system.learning_managment_system_dev.user_managment.Dto.UserCreation;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Entities.Student;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Entities.Teacher;
+import project.learning_managment_system.learning_managment_system_dev.user_managment.KafkaConfig.Producer;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Repositories.Student_Repo;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Repositories.Teacher_Repo;
 import project.learning_managment_system.learning_managment_system_dev.user_managment.Services.ManagementService.ServiceUser;
@@ -15,6 +19,7 @@ import project.learning_managment_system.learning_managment_system_dev.user_mana
 import project.learning_managment_system.learning_managment_system_dev.user_managment.mappers.Implementation.Teacher_Mapper;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,6 +28,9 @@ public class ServiceTeacher implements ServiceUser<Teacher_Dto> {
     public Teacher_Repo teacherRepo;
     @Autowired
     public Teacher_Mapper teacherMapper;
+    @Autowired
+    public Producer producer;
+    public BCryptPasswordEncoder bCrypt=new BCryptPasswordEncoder(12);
 
     @Override
     public List<Teacher_Dto> getUsers() {
@@ -42,17 +50,61 @@ public class ServiceTeacher implements ServiceUser<Teacher_Dto> {
     public Teacher_Dto updateUser(Teacher_Dto user, int id) {
         Teacher teacher =this.teacherRepo.findById(id).orElseThrow(()->new IllegalArgumentException("TEACHER NOT FOUND"));
         this.teacherMapper.updateEntityFromDto(user,teacher);
+        this.producer.syncData(this.mapTo(user));
         return this.teacherMapper.toDto(this.teacherRepo.save(teacher));
     }
 
     @Override
     public void deleteUser(int id) {
-        this.teacherRepo.deleteById(id);
+        Optional<Teacher> teacher=this.teacherRepo.findById(id);
+        if(teacher.isPresent()){
+            this.teacherRepo.deleteById(id);
+            this.producer.syncData(UserCreation.builder()
+                    .mail(teacher.get().getMail())
+                    .operation("DELETE")
+                    .build());
+        }
+        else {
+            throw  new IllegalArgumentException("USER NOT FOUND FOR ID:"+id);
+        }
     }
     @Override
     public Teacher_Dto getMyProfile(Jwt token) {
         JwtExtractor jwtExtractor = new JwtExtractor();
         return this.teacherRepo.findByMail(jwtExtractor.extractClaim(token, "preferred_username"))
                 .map(teacherMapper::toDto).orElseThrow(() -> new IllegalArgumentException("TEACHER NOT FOUND"));
+    }
+
+    @Override
+    public UserCreation mapTo(Teacher_Dto user) {
+        return UserCreation.builder()
+                .mail(user.getMail())
+                .lastname(user.getLastname())
+                .firstname(user.getFirstname())
+                .operation("UPDATE_PROFILE")
+                .build();
+    }
+
+    @Override
+    @Transactional
+    public void deleteMyProfile(Jwt token) {
+        JwtExtractor jwtExtractor=new JwtExtractor();
+        String mail =jwtExtractor.extractClaim(token,"preferred_username");
+        this.teacherRepo.deleteByMail(mail);
+        this.producer.syncData(UserCreation.builder()
+                .operation("DELETE")
+                .mail(mail)
+                .build());
+    }
+
+    @Override
+    public void updatePassword(UserCreation userCreation,int id) {
+        userCreation.setId(id);
+        this.teacherRepo.findById(id)
+                .ifPresent(user->{
+                    user.setPassword(bCrypt.encode(userCreation.getPassword()));
+                    this.teacherRepo.save(user);
+                });
+        this.producer.syncData(userCreation);
     }
 }
